racetrack_client/requirements.txt
PyYAML >= 5.4
pydantic >= 1.9, < 2
backoff >= 1.10
python-socketio >= 5.4
pytz >= 2022.2.1
